refactor: move settings module to the infrastructure layer
